requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
websockets
numpy
Pillow
pyngrok
httpx